pagala project/requirements.txt
streamlit
python-dotenv

# --- LangChain stable versions ---
langchain==0.2.12
langchain-core==0.2.27
langchain-community==0.2.10
langchain-google-genai==1.0.6
langchain-openai==0.1.7

# --- Model Providers ---
openai
google-generativeai

# --- Vector Store / Documents ---
faiss-cpu
pypdf
docx2txt

# --- Embeddings ---
sentence-transformers
